+ classes

+++ src/main/java/com/backend/greekgods/Exercise.kt
@@ -1,5 +1,4 @@
 package com.backend.greekgods
 
-import java.math.BigInteger
 
-class Exercise(var id: BigInteger, var name: String, var sets: String, var reps: String, var weight: String, var description: String) {}
+class Exercise(var id: Long, var name: String, var sets: String, var reps: String, var weight: String, var description: String) {}
